+++ src/Shared/WebLib/src/js/mapView.js
@@ -38,7 +38,7 @@ export class MapView {
         console.groupCollapsed("%cMapView%ccreating new 3D map view", MapView.consoleLogStyle);
         console.time("ctor");
 
-        this.options = Object.assign(options, {
+        this.options = Object.assign({
             id: 'mapElement',
             liveTrackToolbarId: 'liveTrackToolbar',
             initialCenterPoint: { latitude: 47.67, longitude: 11.88 },
@@ -46,7 +46,7 @@ export class MapView {
             hasMouse: false,
             useAsynchronousPrimitives: true,
             useEntityClustering: true
-        });
+        }, options);
 
         this.options.bingMapsApiKey = this.options.bingMapsApiKey || 'AuuY8qZGx-LAeruvajcGMLnudadWlphUWdWb0k6N6lS2QUtURFk3ngCjIXqqFOoe';
         this.options.cesiumIonApiKey = this.options.cesiumIonApiKey || 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJqdGkiOiJiZWMzMjU5NC00MTg4LTQwYmEtYWNhYi01MDYwMWQyZDIxNTUiLCJpZCI6NjM2LCJpYXQiOjE1MjUzNjQ5OTN9.kXik5Mg_-01LBkN-5OTIDpwlMcuE2noRaaHrqjhbaRE';