+++ Changelog.md
@@ -9,6 +9,7 @@ Released on ?.
 - Layers, tracks and locations now have a 3-dot-button to show the context menu
 - The context menus now have icons and are displayed as popup dialogs
 - Updated included Paraglidingspots European Alps locations to version 2.02
+- Display name and description from CZML layers
 - using CesiumJS version 1.92
 
 🐛 Bugfixes