+++ Changelog.md
@@ -10,7 +10,7 @@ Released on ?.
 - The context menus now have icons and are displayed as popup dialogs
 - Updated included Paraglidingspots European Alps locations to version 2.02
 - Display name and description from CZML layers
-- using CesiumJS version 1.92
+- using CesiumJS version 1.95
 
 🐛 Bugfixes